ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.0
torch==2.1.0
pillow==10.1.0
sentence-transformers==2.2.2
numpy==1.26.0
python-multipart==0.0.6
requests==2.31.0
